fix for android crash (#35)
* fix for android crash

For some reason, a recent commit removed the check for a null return
from `getCurrentActivity()`. But since `getCurrentActivity()` might return
null, the app will crash with a NullPointerException.

This restores the null check.

It also fixes inconsistent handling of null `tag` argument in both
iOS and Android. This does not affect usage from JS because the JS
interface will not pass null to that argument; however other native
code could still call these methods and pass null for `tag`.

* remove unnecessary parens

---------

Co-authored-by: Marc Shilling <marcshilling@gmail.com>